Refactoring: moved parameters in own package
